refactor(scraping): ScrapingインターフェースをDomain/Gateways層へ移動
将来的なパースのAI（LLM）移行を見据え、クリーンアーキテクチャおよびDIP（依存関係逆転の原則）を徹底するため、以下のインターフェースをインフラ層からドメイン層へ格上げしました。

- ISuumoHtmlFetcher.cs
- ISuumoHtmlParser.cs

これに伴い、実装クラス（SuumoHtmlFetcher）のusing参照を更新しました。
